평균이 정수로 계산되는 오류 수정

+++ 2학년_2학기/자바/MovieRating.java
@@ -31,7 +31,7 @@ public class MovieRating {
 
             s.nextLine();
 
-            avgRating[i] = sum / text.length;
+            avgRating[i] = sum / (double)text.length;
 
             System.out.println();
         }
